backend/requirements.txt
python-dotenv
python-telegram-bot
flask
